requirements.txt
autocrop==1.2.0
matplotlib==3.5.0
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
skimage==0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
